pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reasonborn"
version = "1.0.0"
description = "ReasonBorn: A Subject-Specific Small Language Model Architecture with Nested CoT"
readme = "README.md"
authors = [{name = "Soham Pal", email = "botmaker583@gmail.com"}]
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = [
    # --- Core ML (ROCm wheels via: pip install --extra-index-url https://download.pytorch.org/whl/rocm6.1) ---
    "torch==2.1.2",
    "transformers==4.35.0",
    "datasets==2.14.0",
    "accelerate==0.24.0",
    "deepspeed==0.11.0",
    "wandb==0.16.0",
    "tensorboard==2.15.0",
    # --- Numerical ---
    "numpy==1.24.3",
    "scipy==1.11.3",
    "scikit-learn==1.3.2",
    "pandas==2.1.3",
    "tqdm==4.66.1",
    # --- Formal Verification ---
    "z3-solver==4.12.2.0",
    "sympy==1.12",
    # --- Vector Search (CPU-based, ROCm-compatible) ---
    "faiss-cpu==1.7.4",
    # --- Embeddings ---
    "sentence-transformers==2.2.2",
    "datasketch>=1.6.0",
    # --- Privacy ---
    "opacus>=1.4.0",
    # --- Serving ---
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "pydantic>=2.4.2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest==7.4.3", "black", "isort", "mypy"]
# AMD MIGraphX for inference optimization (replaces NVIDIA TensorRT)
edge = ["migraphx"]

[tool.setuptools.packages.find]
where = ["src"]
